Move QueryStructure to its own file
